pyproject.toml
[tool.poetry]
name = "video-hybrid-search"
version = "0.1.0"
description = ""
authors = ["Yaman Al-Nabulsi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
weaviate-client = "^3.26.2"
sentence-transformers = "^2.7.0"
streamlit = "^1.33.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
